Evolta main purpsoe is monitoring electrical leaks,
Evolta main purpsoe is monitoring electrical leaks, especially in the case of natural disasters. In these situations, an electrician would make use of a Non-Contact High Voltage Indicator. The issue here is that the power lines are not always safe to approach or are in unreachable locations. Thus, they have developed a drone with an NCHV Indicat…